Data Set - Alcohol
The data explored is contained within the alcohol.txt file. The dataset contains 7 columns and 46 rows. The dataset is recollection of information from various states, especially regarding population and drinking data. The columns are arranged as follows: index, One, urbanPop, lateBirths, wineC, liquorC, and cirrhosis. 
The index column represents a numbering to distinguish different entries within the dataset. 
One is just a column filled with the number 1. There is no true explanation for the need or reason behind this column, it seems counterintuitive. 
urbanPop represents the size of the urban population as a percentage of the total population. 
lateBirths is the reciprocal of the number of births to women between 45 to 49, times a 100. 
wineC is the wine consumption per capita
liquor is the liquor consumption per capita
Finally, cirrhosis is the death rate from Cirrhosis

	Having access to this information, I am interested in noting if a relationship can be established between the death rate from Cirrhosis and the other variables. That way if a linear relationship can be established, then different expectations regarding the death certainty from Cirrhosis can be formed when considering different approaches, including medical ones. Knowing different consumption patterns of alcohol (wine and liquor) and some demographic information (composure of urban population and late births), we could potentially predict what the death rate from having Cirrhosis will be.

Note: an α=0.05 value will be established, in order to assess statistical significance. The α is the probability required threshold (p-Value) of rejecting the null hypothesis.



The dataset comes from the Department of Scientific Computing of Florida State University. The dataset is located within a section of the web page of John Burkardt, who is a visiting research associate at the previously mentioned department. The dataset is contained within a file called x20.txt, but this file contains other strings that serve as the description and references of the data. For this reason, I extracted the raw data from the x20.txt file and copied it to a new file which I named alcohol.txt. 
	The file alcohol.txt includes all of the columns in the original x20.txt file. When I load the data and begin to work on it, I immediate remove 2 columns, as it can be noted in the code.txt. I preferred eliminating these columns, instead of just not including in them in the model for simplicity purposes. The logic behind eliminating these two columns is the following:

	index: This column just represents the index, and other than that it represents no contribution to the data. The data was obtained from different states, and the order in which it was obtained (i.e. the index) does not contribute towards the overall relationship we are trying to find.
	One: This column is just filled with the number 1. Logically, this column makes no sense. It remains constant for all entries in the dataset. The x20.txt file does not offer a compelling reasoning behind why this column is part of the data, nor what its purpose is. Statistically, this column should be removed. I ran an F test of a model without it after noting that logically it made no sense and the p-Value was 1, which means that the model without it should be kept. For these reasons, I just decided to eliminate it from the beginning of the analysis.

I found the data particularly interesting, because the relationship I am trying to establish makes sense intuitively. One should expect for the cirrhosis death rate for a particular state to increase if the population consumption of alcohol increases, that being wine or liquor. With the analysis of this data, I will note if the data follows intuition.
The data relationship that will be explored is going to be established by means of a multivariate linear regression.


Source of data: https://people.sc.fsu.edu/~jburkardt/datasets/regression/x20.txt